third_party/WebKit/LayoutTests/media/video-display-none-crash-expected.txt

CONSOLE ERROR: Uncaught (in promise) AbortError: The play() request was interrupted by a call to pause().
Test that pause() after changing display to "none" doesn't cause a crash.

END OF TEST
